Refactor into plugin structure
