--- variables.php
@@ -1,15 +1,0 @@
-<?php
-
-if (!defined('ABSPATH')) {
-    exit; // Exit if accessed directly
-}
-
-return [
-    "client_id" => "cb36d4b5-602e-408d-bcc8-26fa2c3aaf35",
-    "client_secret" => "a66dab93-5c25-4d68-bd5c-3157d8a55928",
-    "redirect_uri" => "https://staging2.steelmark.com.au/wp-json/hubspot/v1/oauth/callback",
-    "scopes" => "crm.objects.line_items.read crm.objects.line_items.write oauth conversations.read conversations.write crm.objects.contacts.write e-commerce sales-email-read crm.objects.companies.write crm.objects.companies.read crm.objects.deals.read crm.objects.deals.write crm.objects.contacts.read"
-];
-
-?>
-
